moved validate changelog script out
